server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.31.1
python-multipart==0.0.9
asyncpg>=0.30.0
redis==5.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
email-validator==2.2.0
pydantic==2.10.3
pydantic-settings==2.8.1
python-dateutil==2.9.0.post0
httpx==0.27.2
aioredis==2.0.1
orjson==3.10.16
slowapi==0.1.9
limits==3.14.1